Move to proper configuration
