+++ Heroku/.idea/runConfigurations/Backend__OGM_.xml
@@ -2,7 +2,6 @@
   <configuration default="false" name="Backend (OGM)" type="JetRunConfigurationType">
     <option name="MAIN_CLASS_NAME" value="net.twisterrob.cinema.cineworld.backend.MainKt" />
     <module name="Cinema-Heroku.backend.endpoint.main" />
-    <option name="PROGRAM_PARAMETERS" value="-P:environment=production" />
     <shortenClasspath name="NONE" />
     <option name="VM_PARAMETERS" value="-Dlog4j.configurationFile=log4j2.xml,log4j2-endpoint.xml" />
     <method v="2">
