+++ Heroku/backend/endpoint/src/main/kotlin/net/twisterrob/cinema/cineworld/backend/ktor/config.kt
@@ -3,6 +3,7 @@
 package net.twisterrob.cinema.cineworld.backend.ktor
 
 import io.ktor.server.config.ApplicationConfig
+import java.io.File
 
 enum class Env {
 	PRODUCTION,
@@ -10,7 +11,19 @@ enum class Env {
 	DEVELOPMENT
 }
 
+val ApplicationConfig.app: ApplicationConfig
+	get() = this.config("twisterrob.cinema")
+
 val ApplicationConfig.environment: Env
-	get() = this.propertyOrNull("environment")
-		?.let { Env.valueOf(it.getString().uppercase()) }
-		?: Env.DEVELOPMENT
+	get() = this.app.property("environment")
+		.let { Env.valueOf(it.getString().uppercase()) }
+
+val ApplicationConfig.staticRootFolder: File
+	get() = this.app.property("staticRootFolder")
+		.getString()
+		.let(::File)
+
+val ApplicationConfig.fakeRootFolder: File
+	get() = this.app.property("fakeRootFolder")
+		.getString()
+		.let(::File)
