+++ Heroku/backend/endpoint/src/main/resources/application.conf
@@ -9,3 +9,13 @@ ktor {
 		]
 	}
 }
+
+twisterrob {
+	cinema {
+		environment = development
+		# For serving static content (`.` is the Heroku project folder when ran from IDEA).
+		staticRootFolder = ./deploy/frontend/static
+        # Fallback folder for requests for testing (`.` is the Heroku project folder when ran from IDEA).
+		fakeRootFolder = ./backend/src/test/fake
+	}
+}
