+++ Heroku/deploy/appengine/src/main/appengine/app.yaml
@@ -6,8 +6,13 @@ automatic_scaling:
   max_instances: 1
 
 # `static` is a folder in the deployment directory, it's populated by :deploy:appengine:copyFrontendResources.
-entrypoint: java -Xmx64M -jar twisterrob-cinema-backend-endpoint.jar static
-  
+entrypoint: >
+  java
+  -Xmx64M
+  -jar twisterrob-cinema-backend-endpoint.jar
+  -P:twisterrob.cinema.environment=production
+  -P:twisterrob.cinema.staticRootFolder=static
+
 env_variables:
   # Used by endpoint and sync. URL of the Neo4J graph database server, including authentication details. e.g. neo4j+s://user:pass@host:1234",
   NEO4J_URL: @NEO4J_URL@
